checkpoint. shared blue axe knight. most mappings.
