+++ config/splat.us.sttop.yaml
@@ -39,21 +39,24 @@ segments:
     subalign: 4
     subsegments:
       - [0x0, data]
-      - [0x8E4, data, e_init]
-      - [0xA30, data]
+      - [0x8E8, data, e_init]
+      - [0xA98, data, e_breakable]
+      - [0xAF8, data, 296F4]
       - [0xE28, .data, st_update]
       - [0xE74, .data, collision]
       - [0x1234, .data, e_red_door]
-      - [0x124C, data, st_common]
+      - [0x124C, .data, st_common]
       - [0x144C, data, e_collect]
-      - [0x17A0, data, e_misc]
+      - [0x17A0, .data, e_misc]
       - [0x1858, .data, e_particles]
-      - [0x18D8, data, e_room_fg]
-      - [0x1970, data]
-      - [0x19BC, data, e_blue_axe_knight]
+      - [0x18D8, .data, e_room_fg]
+      - [0x1964, data]
+      - [0x19BC, .data, e_blue_axe_knight]
       - [0x1B78, data]
       - [0x288E8, .rodata, e_lock_camera]
       - [0x28904, .rodata, us]
+      - [0x2891C, .rodata, e_breakable]
+      - [0x2893C, .rodata, 296F4]
       - [0x2898C, .rodata, cutscene]
       - [0x28A64, .rodata, e_red_door]
       - [0x28A7C, .rodata, e_collect]
@@ -64,6 +67,8 @@ segments:
       - [0x28C38, c, e_room_bg]
       - [0x28D00, c, e_lock_camera]
       - [0x29250, c, us]
+      - [0x29390, c, e_breakable]
+      - [0x296F4, c, 296F4]
       - [0x2B768, c, cutscene]
       - [0x2D630, c, st_update]
       - [0x2DA64, c, collision]
@@ -80,6 +85,9 @@ segments:
       - [0x38BB8, c, e_blue_axe_knight]
       - [0x3A3C8, c, prim_helpers]
       - [0x3ABF8, c, e_clock_tower]
-      - [0x3B7F0, sbss]
+      - [0x3B7F0, bss]
+      - [0x3B850, .bss, create_entity]
+      - [0x3B860, bss]
+      - [0x3C51C, bss, e_collect]
       # - [0x, sbss, cutscene]
   - [0x3C55C]
