+++ src/st/top/296F4.c
@@ -1,0 +1,26 @@
+#include "common.h"
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_801A8328);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801A99EC);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801A9CB8);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801A9F0C);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AA354);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AA624);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AABA4);
+
+// charal
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AB308);
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AB45C);
+
+// alive
+
+INCLUDE_ASM("st/top/nonmatchings/296F4", func_us_801AB67C);
+
