+++ src/st/top/create_entity.c
@@ -1,33 +1,4 @@
-#include "common.h"
+#include "top.h"
 
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntityFromLayout);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntityWhenInVerticalRange);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntityWhenInHorizontalRange);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", FindFirstEntityToTheRight);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", FindFirstEntityToTheLeft);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntitiesToTheRight);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntitiesToTheLeft);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", FindFirstEntityAbove);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", FindFirstEntityBelow);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntitiesAbove);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntitiesBelow);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", InitRoomEntities);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", UpdateRoomPosition);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntityFromCurrentEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/create_entity", CreateEntityFromEntity);
-
+#include "../pfn_entity_update.h"
+#include "../create_entity.h"
