+++ src/st/top/e_blue_axe_knight.c
@@ -1,20 +1,2 @@
-#include "common.h"
-
-// e_blue_axe_night
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", AxeKnightUnkFunc1);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", func_us_801CA51C);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", func_us_801CAD20);
-
-INCLUDE_RODATA("st/top/nonmatchings/e_blue_axe_knight", D_us_801A8C00);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", EntityAxeKnightBlue);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", EntityAxeKnightRotateAxe);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", EntityAxeKnightThrowingAxe);
-
-INCLUDE_ASM("st/top/nonmatchings/e_blue_axe_knight", func_us_801CBA30);
-
+#include "top.h"
+#include "../e_blue_axe_knight.h"
