+++ src/st/top/e_breakable.c
@@ -1,0 +1,3 @@
+#include "common.h"
+
+INCLUDE_ASM("st/top/nonmatchings/e_breakable", EntityBreakable);
