+++ src/st/top/e_misc.c
@@ -1,37 +1,3 @@
-#include "common.h"
+#include "top.h"
 
-INCLUDE_ASM("st/top/nonmatchings/e_misc", CheckColliderOffsets);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityUnkId13);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityExplosionVariantsSpawner);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityGreyPuffSpawner);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityExplosionVariants);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityGreyPuff);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityOlroxDrool);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", UnkCollisionFunc5);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", UnkCollisionFunc4);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityIntenseExplosion);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", InitializeUnkEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", MakeEntityFromId);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", MakeExplosions);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", EntityBigRedFireball);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", UnkRecursivePrimFunc1);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", UnkRecursivePrimFunc2);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", ClutLerp);
-
-INCLUDE_ASM("st/top/nonmatchings/e_misc", PlaySfxPositional);
+#include "../e_misc.h"
