+++ src/st/top/st_common.c
@@ -1,69 +1,3 @@
-#include "common.h"
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", DestroyEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", DestroyEntitiesFromIndex);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", PreventEntityFromRespawning);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", AnimateEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", UnkAnimFunc);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetDistanceToPlayerX);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetDistanceToPlayerY);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetSideToPlayer);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", MoveEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", FallEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", UnkCollisionFunc3);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", UnkCollisionFunc2);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", AllocEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetSineScaled);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetSine);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", SetEntityVelocityFromAngle);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", Ratan2Shifted);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetAngleBetweenEntitiesShifted);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetAnglePointToEntityShifted);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", AdjustValueWithinThreshold);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", UnkEntityFunc0);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", Ratan2);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetAngleBetweenEntities);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetAnglePointToEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetNormalizedAngle);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", SetStep);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", SetSubStep);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", EntityExplosionSpawn);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", InitializeEntity);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", EntityDummy);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", UnkCollisionFunc);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", CheckFieldCollision);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", GetPlayerCollisionWith);
-
-INCLUDE_ASM("st/top/nonmatchings/st_common", ReplaceBreakableWithItemDrop);
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "top.h"
+#include "../st_common.h"
