+++ src/st/top/us.c
@@ -3,25 +3,4 @@
 
 INCLUDE_ASM("st/top/nonmatchings/us", func_us_801A9250);
 
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801A9390);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_801A8328);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801A99EC);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801A9CB8);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801A9F0C);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AA354);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AA624);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AABA4);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AB308);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AB45C);
-
-INCLUDE_ASM("st/top/nonmatchings/us", func_us_801AB67C);
